requirements.txt
ascii-magic==2.3.0
python-dotenv==1.0.1
requests==2.32.2